Make fund and minAmount fields Int

+++ app/src/main/java/ua/nanit/extop/monitoring/data/Rate.kt
@@ -4,8 +4,8 @@ data class Rate(
     val exchanger: String,
     val amountIn: Float,
     val amountOut: Float,
-    val minAmount: Float,
-    val fund: Float,
+    val minAmount: Int,
+    val fund: Int,
     val link: String,
     val reviewsLink: String?,
 )
